requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-docx>=0.8.11
PyYAML>=6.0
pywebview>=4.2.2
python-multipart>=0.0.6
pydantic>=2.0.0
